pyproject.toml
[project]
name = "autodoctor"
version = "1.0.0"
description = "Home Assistant integration that validates automations to detect state-related issues"
requires-python = ">=3.12"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-homeassistant-custom-component>=0.13.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
